requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.11.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
kaggle==1.5.9
mysqlclient==2.0.3
Pillow==8.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
requests==2.24.0
six==1.15.0
slugify==0.0.1
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
tqdm==4.51.0
urllib3==1.25.11
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
